requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
filelock==3.15.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llvmlite==0.43.0
markdownify==0.11.6
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jobspy==1.1.51
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.32
sympy==1.13.1
threadpoolctl==3.5.0
tls-client==1.0.1
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.10.0
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.6
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
